deprecate DeviceApi class now fully handled in helper package
